Moved tests from ASqlServer.Test to SqlServerTests
