+++ src/ASqlServer.Test/GlobalUsings.cs
@@ -1,25 +1,11 @@
-global using ABulkCopy.ASqlServer;
-global using ABulkCopy.ASqlServer.Column;
 global using ABulkCopy.ASqlServer.Column.ColumnTypes;
 global using ABulkCopy.Common;
-global using ABulkCopy.Common.Config;
-global using ABulkCopy.Common.Database;
-global using ABulkCopy.Common.Extensions;
-global using ABulkCopy.Common.Reader;
-global using ABulkCopy.Common.TestData;
-global using ABulkCopy.Common.Types;
 global using ABulkCopy.Common.Types.Column;
-global using ABulkCopy.Common.Types.Index;
 global using ABulkCopy.Common.Types.Table;
-global using ABulkCopy.Common.Utils;
 global using ABulkCopy.Common.Writer;
 global using FluentAssertions;
-global using Microsoft.Extensions.Configuration;
-global using Moq;
 global using Serilog;
-global using System.Data;
 global using System.IO.Abstractions.TestingHelpers;
 global using Testing.Shared;
 global using Testing.Shared.SqlServer;
 global using Xunit;
-global using Xunit.Abstractions;
