+++ src/SqlServerTests/MssDataWriterTestsForDateTime.cs
@@ -1,4 +1,4 @@
-﻿namespace ASqlServer.Test;
+﻿namespace SqlServerTests;
 
 public class MssDataWriterTestsForDateTime : MssDataWriterTestBase
 {
