+++ src/SqlServerTests/MssDataWriterTestsMisc.cs
@@ -1,4 +1,4 @@
-﻿namespace ASqlServer.Test;
+﻿namespace SqlServerTests;
 
 public class MssDataWriterTestsMisc : MssDataWriterTestBase
 {
