+++ src/SqlServerTests/MssSystemTablesTests.cs
@@ -1,4 +1,4 @@
-namespace ASqlServer.Test;
+namespace SqlServerTests;
 
 public class MssSystemTablesTests : MssTestBase
 {
